xeno_backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Django==4.2.20
django-cors-headers==4.7.0
djangorestframework==3.16.0
fonttools==4.57.0
idna==3.10
joblib==1.4.2
jsonpickle==4.0.5
kiwisolver==1.4.8
kombu==5.5.2
matplotlib==3.10.1
more-itertools==10.6.0
music21==9.5.0
numpy==1.26.4
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
webcolors==24.11.1